+++ src/main/java/roomescape/controller/reservation/ReservationController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.reservation;
 
 import java.util.HashMap;
 import java.util.List;
@@ -14,9 +14,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.Reservation;
-import roomescape.controller.request.ReservationRequest;
-import roomescape.controller.response.ReservationResponse;
+import roomescape.controller.reservation.request.ReservationRequest;
+import roomescape.controller.reservation.response.ReservationResponse;
+import roomescape.model.Reservation;
 
 @RequestMapping("/reservations")
 @RestController
