+++ src/main/java/roomescape/model/Time.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.model;
 
 import java.time.LocalTime;
